@@ -17,45 +17,51 @@ final darkTextStyle = kRobotoTextStyle.copyWith(
 final lightTheme = ThemeData(
   primaryColor: kLightPrimaryColor,
   shadowColor: kLightPrimaryColor,
-  backgroundColor: Colors.white,
+  colorScheme: ColorScheme.light(
+    primary: kLightPrimaryColor,
+    secondary: Colors.white,
+  ),
   visualDensity: VisualDensity.adaptivePlatformDensity,
   textTheme: TextTheme(
-    headline5: lightTextStyle.copyWith(
+    headlineSmall: lightTextStyle.copyWith(
       fontSize: 22,
       fontWeight: FontWeight.bold,
     ),
-    headline6: lightTextStyle.copyWith(
+    titleLarge: lightTextStyle.copyWith(
       fontSize: 17,
       fontWeight: FontWeight.bold,
     ),
-    subtitle1: lightTextStyle.copyWith(
+    titleMedium: lightTextStyle.copyWith(
       fontSize: 13,
     ),
-    bodyText1: lightTextStyle.copyWith(
+    bodyLarge: lightTextStyle.copyWith(
       fontSize: 20,
     ),
   ),
 );
 
 final darkTheme = ThemeData(
-  primaryColor: Colors.black,
-  backgroundColor: Colors.black87,
   shadowColor: Colors.white,
+  colorScheme: ColorScheme.dark(
+    primary: Colors.black,
+    secondary: Colors.white,
+    background: Colors.black87,
+  ),
   cardColor: kLightPrimaryColor.withOpacity(0.5),
   visualDensity: VisualDensity.adaptivePlatformDensity,
   textTheme: TextTheme(
-    headline5: darkTextStyle.copyWith(
+    headlineSmall: darkTextStyle.copyWith(
       fontSize: 22,
       fontWeight: FontWeight.bold,
     ),
-    headline6: darkTextStyle.copyWith(
+    titleLarge: darkTextStyle.copyWith(
       fontSize: 17,
       fontWeight: FontWeight.bold,
     ),
-    subtitle1: darkTextStyle.copyWith(
+    titleMedium: darkTextStyle.copyWith(
       fontSize: 13,
     ),
-    bodyText1: darkTextStyle.copyWith(
+    bodyLarge: darkTextStyle.copyWith(
       fontSize: 22,
     ),
   ),
